Fixed user params; removed destroy work for non-users

+++ app/controllers/works_controller.rb
@@ -15,7 +15,6 @@ class WorksController < ApplicationController
   
   def create 
     if session[:user_id]
-      raise
       @work = Work.new(work_params)
       if @work.save
         flash[:success] = "Successfully created " + @work.category + " " + @work.id.to_s
@@ -35,7 +34,6 @@ class WorksController < ApplicationController
   
   def update
     if session[:user_id]
-      raise
       if @work.update(work_params)
         redirect_to work_path
       else
@@ -49,11 +47,16 @@ class WorksController < ApplicationController
   end
   
   def destroy
-    votes = Vote.where(work_id: @work.id)
-    votes.destroy_all
-    @work.destroy
-    flash[:success] = "Successfully destroyed " + @work.category + " " + @work.id.to_s
-    redirect_to root_path
+    if session[:user_id]
+      votes = Vote.where(work_id: @work.id)
+      votes.destroy_all
+      @work.destroy
+      flash[:success] = "Successfully destroyed " + @work.category + " " + @work.id.to_s
+      redirect_to root_path
+    else
+      flash[:error] = "You must log in to do that"
+      redirect_to works_path
+    end
   end
   
   private
